Multiple changes
Fixed the final stop so that all players will come to a stop once they hit the finish gate.
Changed the code now so if there are 3 players and the 2nd leaves, the other 2 players can still play (as of currently player 3 will still be on lane 3)
Changed the countdown and game manager script to support the change
More scripts will have to be changed but ill leave that for the rest of the programming team.