Improved build tool, fixed bathtub and new winner menu
Improved the build tool so when the user clicks cancel it will no longer build to an emptry directory. The new bathtub scaling issue should be fixed now both on the server and the client. Added a mock-up winner UI; shows the player name, position and time. Works dynamically with different amounts of players.